feat: add code splitting setup
