+++ src/screens/about.js
@@ -1,0 +1,14 @@
+import React from 'react'
+
+import { Link } from "react-router-dom"
+
+
+const About = () => (
+  <div>
+    <h2>About</h2>
+<Link to="/">home</Link>
+
+  </div>
+);
+
+export default About
